backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
tavily-python==0.5.0
openai==1.51.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
playwright==1.48.0
websockets==13.0
